# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the swagger of Yet Another Issue Tracker"
  version: "1.0.0"
  title: "Swagger YAIT"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "fabio.percivaldi@hotmail.it"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/v1"
...
schemes:
- "https"
- "http"
security:
- ApiKeyAuth: []
paths:
...
  /projects:
    post:
      tags:
      - "Project"
      summary: "Add a new project"
      description: ""
      operationId: "addProject"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Sprint object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/Project"
        x-exportParamName: "Body"
      responses:
        "201":
          description: "Created"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
        "405":
          description: "Invalid input"
securityDefinitions:
  ApiKeyAuth:
    type: "apiKey"
    name: "X-API-KEY"
    in: "header"
definitions:
...
  Project:
    type: "object"
    properties:
      id:
        type: "string"
      client:
        type: "string"
      type:
        type: "string"
    example:
      client: "client"
      id: "id"
      type: "type"
...