# go/api_project.go
/*
 * Swagger YAIT
 *
 * This is the swagger of Yet Another Issue Tracker
 *
 * API version: 1.0.0
 * Contact: fabio.percivaldi@hotmail.it
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

import (
	"encoding/json"
	"errors"
	"log"
	"net/http"
)

var (
	errGeneric = errors.New("internal Server Error")
	// errBadRequest = errors.New("bad Request")
)

func createProject() ([]byte, error) {
	response := CreateProjectResponse{
		Id: "123",
	}
	responseBody, err := json.Marshal(response)
	if err != nil {
		return nil, err
	}

	return responseBody, nil
}

func AddProject(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusOK)
	projectId, err := createProject()

	if err != nil {
		log.Fatalln("failed response unmarshalling")
		http.Error(w, errGeneric.Error(), http.StatusInternalServerError)
		return
	}

	w.Write(projectId)
}