# go/model_project.go
/*
 * Swagger YAIT
 *
 * This is the swagger of Yet Another Issue Tracker
 *
 * API version: 1.0.0
 * Contact: fabio.percivaldi@hotmail.it
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Project struct {
	Id string `json:"id,omitempty"`

	Client string `json:"client,omitempty"`

	Type_ string `json:"type,omitempty"`
}

type CreateProjectResponse struct {
	Id string `json:"id,omitempty"`
}